Split the tests into separate files
